+++ addOns/paramdigger/CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update minimum ZAP version to 2.12.0.
 - Maintenance changes.
-
+- Default number of threads to 2 * processor count.
 
 ## [0.1.0] - 2022-08-22
 ### Changed