+++ addOns/spiderAjax/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Maintenance changes.
+- Default number of threads to 2 * processor count.
 
 ## [23.10.0] - 2022-10-27
 ### Changed